Can open key file, but not reading properly yet.
Trying to get the first line of the key file to be properly read, but
only the last three lines are being read correctly. The first line
appears to be a combination of all four lines. It doesn't matter if the
file lines are being read with fgets or fscanf.